Modified global config file structure
