+++ pkg/utils/config/config_test.go
@@ -11,7 +11,7 @@
 //    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 //    License for the specific language governing permissions and limitations
 //    under the License.
-package utils
+package config
 
 import (
 	"testing"
@@ -156,3 +156,4 @@ func TestOpensdsConfig(t *testing.T) {
 		t.Error("Test OsdsLet.ApiEndpoint error")
 	}
 }
+
