+++ pkg/utils/config/flag_test.go
@@ -12,7 +12,7 @@
 //    License for the specific language governing permissions and limitations
 //    under the License.
 
-package utils
+package config
 
 import (
 	gflag "flag"
